temp commit

+++ src/main/java/com/tradingBot/model/OptionGreeks.java
@@ -4,11 +4,13 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class OptionGreeks {
-    private Double delta;
-    private Double gamma;
+    private BigDecimal delta;
+    private BigDecimal gamma;
     private Double theta;
     private Double vega;
     private Double rho;
